feat: added analysis levels and setting up components
In this part, we're calculating risks by levels and setting alarms to alert the user about them. This means that we setup components like HC-SR04, a buzzer. Furthermore, we improved the loop thing, using Arduino "milis" function instead using "delay". This change came out of necessity of keep the buzzer ringing (due to a risk) while the system is awaiting for the new data load from the script.